fix: make image not inline
