+++ examples/editor/src/plugins/customTransformers.tsx
@@ -9,8 +9,8 @@ import {
   $createYouTubeNode,
   $isYouTubeNode,
 } from "../../../../packages/milkup-youtube/src/YoutubeNode";
-import { ImageNode } from "../../../../packages/milkup-image/ImageNode";
-import { $createImageNode } from "../../../../packages/milkup-image/ImageNode";
+import { ImageNode } from "../../../../packages/milkup-image/src/ImageNode";
+import { $createImageNode } from "../../../../packages/milkup-image/src/ImageNode";
 
 /* Horizontal line transformers. */
 
@@ -80,32 +80,3 @@ export const YOUTUBE: ElementTransformer = {
 export function $isImageNode(node: LexicalNode | null | undefined): node is ImageNode {
   return node instanceof ImageNode;
 }
-
-export const IMAGE: ElementTransformer = {
-  dependencies: [ImageNode],
-  export: (node: LexicalNode) => {
-    if ($isImageNode(node)) {
-      console.log("got here");
-      // You can choose to include width and height information in a custom manner if needed.
-      return `![${node.__altText}](${node.__src})`;
-    }
-    console.log("asfjdksflk");
-    return null;
-  },
-
-  // This regular expression matches the markdown image syntax.
-  // For example: ![alt text](https://example.com/image.png)
-  regExp: /!\[([^\]]*)\]\((https?:\/\/[^\s)]+)\)/,
-  replace: (parentNode, _props, match, isImport) => {
-    const altText = match[1];
-    const src = match[2];
-    const imageNode = $createImageNode({ src, altText });
-    if (isImport || parentNode.getNextSibling() != null) {
-      parentNode.replace(imageNode);
-    } else {
-      parentNode.insertBefore(imageNode);
-    }
-  },
-
-  type: "element",
-};
