+++ packages/milkup-image/src/index.ts
@@ -1,0 +1,4 @@
+export * from "./ImageComponent";
+export * from "./ImageNode";
+export * from "./transformers";
+export * from "./ImageToolBarPlugin"
